Only register required RESTEasy providers in REST Client
Fixes #707

Moved provider lookup and discovery into `resteasy-common`.
